Refactor ocsp_updater into a separate package
This pulls most of ocsp_updater into its own package. This follows more
closely the pattern we use elsewhere, where we have a struct for the
service, and a main.go that sets it up.
