+++ ocsp_updater/config/ocsp_updater_config.go
@@ -1,0 +1,26 @@
+package ocsp_updater_config
+
+import "github.com/letsencrypt/boulder/cmd"
+
+// Config provides the various window tick times and batch sizes needed
+// for the OCSP updater
+type Config struct {
+	cmd.ServiceConfig
+	DB         cmd.DBConfig
+	ReadOnlyDB cmd.DBConfig
+
+	OldOCSPWindow    cmd.ConfigDuration
+	OldOCSPBatchSize int
+
+	OCSPMinTimeToExpiry          cmd.ConfigDuration
+	ParallelGenerateOCSPRequests int
+
+	SignFailureBackoffFactor float64
+	SignFailureBackoffMax    cmd.ConfigDuration
+
+	SerialSuffixShards string
+
+	OCSPGeneratorService *cmd.GRPCClientConfig
+
+	Features map[string]bool
+}
